requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0

# Optional: for better performance
tensorboard>=2.10.0
